[WIP] !!! Update compoents
Add separate view file for rooting in js stuff, to be further used in router.

React expects capitalized component names. r/reactify-component can
rename with use of :displayName metadata. React router must be provided
capitalized components.

Way it is coded in this commit it does not work. Use `with-meta` or
??add properly metadata to functions??